be/requirements.txt
elasticsearch==7.13.4
flask==3.1.0
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
openai==1.77.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-core==2.27.1
pyjwt==2.10.1
python-dotenv==1.1.0
pytz==2024.1
requests==2.32.3
werkzeug==3.1.3
gunicorn==21.2.0
pypdf==6.6.1
flask-cors==6.0.1
